refactor(cli/audio): split output submodule into file
